Change logic for storaging files.
