+++ .gitignore
@@ -362,3 +362,4 @@ MigrationBackup/
 # Fody - auto-generated XML schema
 FodyWeavers.xsd
 /Provis.WebApi/wwwroot/images
+/Provis.WebApi/wwwroot/attachments
