+++ Provis.Core/Helpers/FileSettings.cs
@@ -2,6 +2,7 @@
 {
     public class FileSettings
     {
+        public bool AllowStoreInAzureBlobStore { get; set; }
         public bool AllowCreateFolderPath { get; set; }
         public bool AllowChangeName { get; set; }
     }
