+++ Provis.Core/Helpers/StorageTypes.cs
@@ -1,0 +1,8 @@
+﻿namespace Provis.Core.Helpers
+{
+    public enum StorageTypes
+    {
+        AzureBlob,
+        Locale
+    }
+}
