+++ Provis.Core/Interfaces/Services/IAzureBlobStorageService.cs
@@ -1,0 +1,6 @@
+﻿namespace Provis.Core.Interfaces.Services
+{
+    public interface IAzureBlobStorageService: IFileService, ICreateDirectory
+    {
+    }
+}
