+++ Provis.Core/Interfaces/Services/ICreateDirectory.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Provis.Core.Interfaces.Services
+{
+    public interface ICreateDirectory
+    {
+        Task CreateDirectoryAsync(string folderPath);
+    }
+}
