+++ Provis.Core/Interfaces/Services/IFileService.cs
@@ -9,6 +9,5 @@ namespace Provis.Core.Interfaces.Services
         Task<string> AddFileAsync(Stream stream, string folderPath, string fileName);
         Task<DownloadFile> GetFileAsync(string path);
         Task DeleteFileAsync(string path);
-        Task CreateDirectoryAsync(string folderPath);
     }
 }
