+++ Provis.Core/Interfaces/Services/ILocaleStorageService.cs
@@ -1,0 +1,7 @@
+﻿namespace Provis.Core.Interfaces.Services
+{
+    public interface ILocaleStorageService: IFileService, ICreateDirectory
+    {
+
+    }
+}
