+++ Provis.Core/StartupSetup.cs
@@ -30,20 +30,13 @@ namespace Provis.Core
 
         public static void AddFileService(this IServiceCollection services, IConfiguration configuration)
         {
-            var allowStoreInAzureBlobStore = configuration.GetSection("FileSettings")
-                .GetValue<bool>("AllowStoreInAzureBlobStore");
+            services.AddScoped(_ =>
+                new BlobServiceClient(configuration.GetSection("AzureBlobStorageSettings")
+                    .GetValue<string>("AccessKey")));
 
-            if (allowStoreInAzureBlobStore)
-            {
-                services.AddScoped(_ =>
-                    new BlobServiceClient(configuration.GetSection("AzureBlobStorageSettings")
-                        .GetValue<string>("AccessKey")));
-                services.AddScoped<IFileService, AzureBlobStorageService>();
-            }
-            else
-            {
-                services.AddScoped<IFileService, FileService>();
-            }
+            services.AddScoped<IFileService, FileService>();
+            services.AddScoped<IAzureBlobStorageService, AzureBlobStorageService>();
+            services.AddScoped<ILocaleStorageService, LocaleStorageService>();
         }
 
         public static void AddFluentValitation(this IServiceCollection services)
@@ -102,6 +95,6 @@ namespace Provis.Core
         public static void ConfigureRolesAccess(this IServiceCollection services, IConfiguration configuration)
         {
             services.Configure<RoleAccess>(configuration.GetSection("RolesAccess"));
-        }        
+        }
     }
 }
